RFC: keep 'readonly' setting
`:edit` seems to reset this always.

This helps when writing a non-user buffer often, but Vim will still ask
about confirmation due to file permissions anyway [1].
Therefore the best seems to use `:w!` then right away, which also makes
this patch unnecessary.

But it might still be sensible to keep the `'readonly'` setting (when it
was changed).

1: https://github.com/vim/vim/commit/5386a123f (patch 7.1.017, have not
   checked current code)