BUG: Regression: Plugin also triggers on writable read-only files (after 166cdb9)
I frequently use :view instead of :edit to open a file as read-only, but then might decide to start editing, and can do so with a simple :w! (which resets the 'readonly' flag).
After updating SudoEdit, the plugin now kicks in (repeatedly, on each write) even when I have full access control to the file!

I think that was not the intention of #64. This adds a check for writability to the added FileChangedRO event handler and skips the plugin activation if the file can be written to.

(This corrects commit 166cdb98c89518b740a674ef117eb3ee68944698)